backend/requirements.txt
Flask
Flask-CORS
gunicorn
cassandra-driver
